chore: 로그인 유저 리뷰 좋아요,싫어요 여부 조회 추가

+++ src/main/java/com/reviewing/review/review/controller/ReviewController.java
@@ -7,6 +7,7 @@ import com.reviewing.review.review.service.ReviewService;
 import com.reviewing.review.review.service.S3Service;
 import jakarta.servlet.http.HttpServletRequest;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
@@ -61,55 +63,58 @@ public class ReviewController {
 
     @GetMapping("/{courseId}")
     public ResponseEntity<List<ReviewResponseDto>> findReviewsByCourse(
-            @PathVariable Long courseId) {
+            @PathVariable Long courseId, HttpServletRequest request) {
 
-        List<ReviewResponseDto> reviews = reviewService.findReviewsByCourse(courseId);
+        List<ReviewResponseDto> reviews;
+
+        String jwtHeader = request.getHeader("Authorization");
+
+        if (jwtHeader != null && jwtHeader.startsWith("Bearer ")) {
+            String token = jwtHeader.replace("Bearer ", "");
+            Long memberId = jwtTokenProvider.getMemberIdByRefreshToken(token);
+
+            reviews = reviewService.findReviewsWithLikedAndDislikedByCourse(courseId, memberId);
+
+            return ResponseEntity.ok().body(reviews);
+        }
+
+        reviews = reviewService.findReviewsByCourse(courseId);
 
         return ResponseEntity.ok().body(reviews);
     }
 
     @PostMapping("/{reviewId}/like")
     public void createReviewLike(@PathVariable Long reviewId,
+            @RequestParam(value = "liked") boolean liked,
             HttpServletRequest request) {
 
         String jwtHeader = request.getHeader("Authorization");
         String token = jwtHeader.replace("Bearer ", "");
         Long memberId = jwtTokenProvider.getMemberIdByRefreshToken(token);
+
+        if (liked) {
+            reviewService.removeReviewLike(reviewId, memberId);
+            return;
+        }
 
         reviewService.createReviewLike(reviewId, memberId);
     }
 
-    @DeleteMapping("/{reviewId}/like")
-    public void removeReviewLike(@PathVariable Long reviewId,
-            HttpServletRequest request) {
-
-        String jwtHeader = request.getHeader("Authorization");
-        String token = jwtHeader.replace("Bearer ", "");
-        Long memberId = jwtTokenProvider.getMemberIdByRefreshToken(token);
-
-        reviewService.removeReviewLike(reviewId, memberId);
-    }
-
     @PostMapping("/{reviewId}/dislike")
     public void createReviewDislike(@PathVariable Long reviewId,
+            @RequestParam(value = "disliked") boolean disliked,
             HttpServletRequest request) {
 
         String jwtHeader = request.getHeader("Authorization");
         String token = jwtHeader.replace("Bearer ", "");
         Long memberId = jwtTokenProvider.getMemberIdByRefreshToken(token);
+
+        if (disliked) {
+            reviewService.removeReviewDislike(reviewId, memberId);
+            return;
+        }
 
         reviewService.createReviewDislike(reviewId, memberId);
-    }
-
-    @DeleteMapping("/{reviewId}/dislike")
-    public void removeReviewDislike(@PathVariable Long reviewId,
-            HttpServletRequest request) {
-
-        String jwtHeader = request.getHeader("Authorization");
-        String token = jwtHeader.replace("Bearer ", "");
-        Long memberId = jwtTokenProvider.getMemberIdByRefreshToken(token);
-
-        reviewService.removeReviewDislike(reviewId, memberId);
     }
 
 }
